Spell UI buttons cleanup and documentation.
